ID Changes
Changed ISBN type to String

+++ Geek-Text/src/main/java/com/example/geektext/rest/Repo/AuthorRepo.java
@@ -2,5 +2,5 @@ package com.example.geektext.rest.Repo;
 import com.example.geektext.rest.Models.Author;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthorRepo extends JpaRepository<Author, Long> {
+public interface AuthorRepo extends JpaRepository<Author, Integer> {
 }
